pyproject.toml
[project]
name = "docagent"
version = "0.1.0"
description = "LLM agent for discovering and assigning topics to documents"
authors = [
    {name = "Keith_McNulty",email = "keith.mcnulty@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic (>=0.54.0,<0.55.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

